requirements.txt
Flask==3.0.0
Flask-Cors==4.0.0
azure-storage-blob==12.19.1
azure-identity==1.15.0
pandas==2.2.1
scikit-learn==1.4.1.post1

